simple coffee fix

+++ src/static/addCoffee.py
@@ -3,7 +3,7 @@ import time
 
 
 name = input("Name: ")
-cups = (int)(input("Cup(s): "))
+cups = (input("Cup(s): "))
 hour = (time.time() - 1618660800) // 3600
 
 d = db.db(host = '192.168.1.100')
